IntroAgentesIA/requirements.txt
python-dotenv
transformers
torch
accelerate
